update typed sample
